backend/pyproject.toml
[tool.poetry]
name = "ai-memoria-backend"
version = "0.1.0"
description = "AI-powered memory management system"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
weaviate-client = "^4.5.3"
python-dotenv = "^1.0.1"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
aiosqlite = "^0.20.0"
openai = "^1.14.2"
anthropic = "^0.18.1"
google-generativeai = "^0.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true 